+++ clap_builder/src/parser/validator.rs
@@ -256,11 +256,13 @@ impl<'cmd> Validator<'cmd> {
                 }
             } else if let Some(group) = self.cmd.find_group(arg_or_group) {
                 debug!("Validator::validate_required:iter: This is a group");
-                if !self
-                    .cmd
-                    .unroll_args_in_group(&group.id)
-                    .iter()
-                    .any(|a| matcher.check_explicit(a, &ArgPredicate::IsPresent))
+                if !is_exclusive_present
+                    && !self.is_missing_required_group_ok(group, conflicts)
+                    && !self
+                        .cmd
+                        .unroll_args_in_group(&group.id)
+                        .iter()
+                        .any(|a| matcher.check_explicit(a, &ArgPredicate::IsPresent))
                 {
                     debug!(
                         "Validator::validate_required:iter: Missing {:?}",
@@ -355,6 +357,11 @@ impl<'cmd> Validator<'cmd> {
             }
         }
         false
+    }
+
+    fn is_missing_required_group_ok(&self, g: &ArgGroup, conflicts: &Conflicts) -> bool {
+        debug!("Validator::is_missing_required_group_ok: {}", g.get_id());
+        !conflicts.gather_conflicts(self.cmd, &g.id).is_empty()
     }
 
     // Failing a required unless means, the arg's "unless" wasn't present, and neither were they